Fix some bugs: - Create new user must be active - Forget-password error, change to english (invalid token) - Confirm when delete account - Fix calculator vacation date - Reson field in vacation should be placeholder - Auto add 12 days for new user - Change remaining day to hours

+++ VacationTracking/Vacation.Web/src/app/admin/user.controller.js
@@ -106,10 +106,12 @@
         };
 
         $scope.deleteUser = function (user, index) {
-            UserService.Delete(user.Id).then(function () {
-                $scope.usersPaging.splice(index, 1);
-                $scope.totalItems--;
-            })
+            if (confirm('Are you sure to delete No.' + (index + 1))) {
+                UserService.Delete(user.Id).then(function () {
+                    $scope.usersPaging.splice(index, 1);
+                    $scope.totalItems--;
+                })
+            }
         }
 
         $scope.detailUser = function (user) {
